requirements.txt
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djongo==1.3.6
psycopg2==2.9.3
pymongo==3.12.3
pytz==2022.2.1
sqlparse==0.2.4
typing_extensions==4.3.0
